requirements.txt
numpy==1.26.4
opencv-python
pandas
scikit-learn
matplotlib
tensorflow
gradio>=4.0
transformers>=4.38.0
torch>=2.0
Pillow